requirements.txt
streamlit
fastapi[all]
python-jose[cryptography]
passlib[bcrypt]
pyjwt
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pytest
httpx
scikit-learn
nltk
pymorphy3
num2words